requirements.txt
asgiref==3.8.1
black==24.10.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
Django==5.1.3
django-extensions==3.2.3
django-shell-plus==1.1.7
mypy-extensions==1.0.0
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pathspec==0.12.1
piexif==1.1.3
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
pyexiv2==2.15.3
python-dotenv==1.0.1
PyWavelets==1.7.0
scipy==1.14.1
sqlparse==0.5.1
tomli==2.1.0
typing_extensions==4.12.2
